Refactor classes into packages to seprate import from file and catalog
